added spring security

+++ src/main/java/pl/krzywyyy/barter/model/dto/OfferDTO.java
@@ -7,8 +7,8 @@ import java.util.Date;
 @Data
 public class OfferDTO
 {
-	private long offeredProductId;
-	private long aimedProductId;
+	private int offeredProductId;
+	private int aimedProductId;
 	private Date offerDate;
 	private Date confirmDate;
 }
